refactor(bkn-trace): remove legacy community provenance routes
