+++ bkn-trace/agent-observability/docs/swagger/docs.go
@@ -54,944 +54,6 @@ const docTemplate = `{
                 }
             }
         },
-        "/business-provenance/conversations": {
-            "get": {
-                "description": "Returns one authorized business-provenance row per conversation, aggregated from real interactions and requests.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "business-provenance"
-                ],
-                "summary": "List observable business conversations",
-                "parameters": [
-                    {
-                        "type": "integer",
-                        "description": "Page number, starting at 1",
-                        "name": "page",
-                        "in": "query"
-                    },
-                    {
-                        "type": "integer",
-                        "description": "Page size, 1..200",
-                        "name": "page_size",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Started at or after this RFC3339 timestamp",
-                        "name": "from",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Started at or before this RFC3339 timestamp",
-                        "name": "to",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Execution status",
-                        "name": "status",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Agent or application",
-                        "name": "agent_or_app",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Exact producing service",
-                        "name": "service",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Exact root tool",
-                        "name": "tool",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Case-insensitive error text",
-                        "name": "error_keyword",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Business domain",
-                        "name": "business_domain",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Knowledge network",
-                        "name": "knowledge_network",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Evidence completeness",
-                        "name": "evidence_completeness",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Question, result, ID, business ref, or error keyword",
-                        "name": "keyword",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.ConversationSummaryPage"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "401": {
-                        "description": "Unauthorized",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/evidence-nodes/{node_id}": {
-            "get": {
-                "description": "Returns one visible claim, evidence ref, or business ref node scoped by trace_id or request_id.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "business-provenance"
-                ],
-                "summary": "Get evidence node details",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "Evidence node ID, for example claim:claim_001",
-                        "name": "node_id",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "type": "string",
-                        "description": "Trace ID scope",
-                        "name": "trace_id",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "BKN request ID scope",
-                        "name": "request_id",
-                        "in": "query"
-                    },
-                    {
-                        "type": "integer",
-                        "description": "Maximum evidence trace batches to read, 1..1000",
-                        "name": "limit",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.EvidenceNodeResponse"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "404": {
-                        "description": "Not Found",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "405": {
-                        "description": "Method Not Allowed",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/interactions": {
-            "get": {
-                "description": "Returns one authorized business-provenance row per interaction, aggregated from real requests.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "business-provenance"
-                ],
-                "summary": "List observable business interactions",
-                "parameters": [
-                    {
-                        "type": "integer",
-                        "description": "Page number, starting at 1",
-                        "name": "page",
-                        "in": "query"
-                    },
-                    {
-                        "type": "integer",
-                        "description": "Page size, 1..200",
-                        "name": "page_size",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Conversation ID",
-                        "name": "conversation_id",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Started at or after this RFC3339 timestamp",
-                        "name": "from",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Started at or before this RFC3339 timestamp",
-                        "name": "to",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Execution status",
-                        "name": "status",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Agent or application",
-                        "name": "agent_or_app",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Business domain",
-                        "name": "business_domain",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Knowledge network",
-                        "name": "knowledge_network",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Evidence completeness",
-                        "name": "evidence_completeness",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Question, result, ID, business ref, or error keyword",
-                        "name": "keyword",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.InteractionSummaryPage"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "401": {
-                        "description": "Unauthorized",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/interactions/{interaction_id}": {
-            "get": {
-                "description": "Aggregates every authorized request and trace in one caller-owned interaction.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "interactions"
-                ],
-                "summary": "Get one observable business interaction",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "BKN interaction ID",
-                        "name": "interaction_id",
-                        "in": "path",
-                        "required": true
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.InteractionSummary"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "401": {
-                        "description": "Unauthorized",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "404": {
-                        "description": "Not Found",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/requests": {
-            "get": {
-                "description": "Returns stable request summaries generated from authorized evidence and artifacts.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "requests"
-                ],
-                "summary": "List observable business requests",
-                "parameters": [
-                    {
-                        "type": "integer",
-                        "description": "Page number, starting at 1",
-                        "name": "page",
-                        "in": "query"
-                    },
-                    {
-                        "type": "integer",
-                        "description": "Page size, 1..200",
-                        "name": "page_size",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Caller-owned conversation ID",
-                        "name": "conversation_id",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "One user interaction ID",
-                        "name": "interaction_id",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Started at or after this RFC3339 timestamp",
-                        "name": "from",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Started at or before this RFC3339 timestamp",
-                        "name": "to",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Execution status",
-                        "name": "status",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Agent or application",
-                        "name": "agent_or_app",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Business domain",
-                        "name": "business_domain",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Knowledge network",
-                        "name": "knowledge_network",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Evidence completeness",
-                        "name": "evidence_completeness",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Question, result, ID, business ref, or error keyword",
-                        "name": "keyword",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.RequestSummaryPage"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "401": {
-                        "description": "Unauthorized",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/requests/{request_id}": {
-            "get": {
-                "description": "Returns an authorized request summary and its business-content availability.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "requests"
-                ],
-                "summary": "Get one observable business request",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "BKN request ID",
-                        "name": "request_id",
-                        "in": "path",
-                        "required": true
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.RequestSummary"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "401": {
-                        "description": "Unauthorized",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "404": {
-                        "description": "Not Found",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/requests/{request_id}/business-graph": {
-            "get": {
-                "description": "Returns claim and business semantic nodes/edges derived from business.refs.resolved events.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "evidence"
-                ],
-                "summary": "Get business semantic graph by BKN request ID",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "BKN request ID",
-                        "name": "request_id",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "type": "integer",
-                        "description": "Maximum evidence trace batches to read, 1..1000",
-                        "name": "limit",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.BusinessGraphResponse"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "404": {
-                        "description": "Not Found",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "405": {
-                        "description": "Method Not Allowed",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/requests/{request_id}/evidence-chain": {
-            "get": {
-                "description": "Returns normalized claim, evidence refs, business refs, pagination, partial reasons, and visibility summary for a request.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "evidence"
-                ],
-                "summary": "Get evidence chain by BKN request ID",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "BKN request ID",
-                        "name": "request_id",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "type": "integer",
-                        "description": "Maximum evidence trace batches to read, 1..1000",
-                        "name": "limit",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.EvidenceChainResponse"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "404": {
-                        "description": "Not Found",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "405": {
-                        "description": "Method Not Allowed",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/requests/{request_id}/snapshot-preview": {
-            "get": {
-                "description": "Returns a metadata-only governed snapshot manifest preview without creating or exposing object storage locations.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "evidence"
-                ],
-                "summary": "Get evidence snapshot preview by BKN request ID",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "BKN request ID",
-                        "name": "request_id",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "type": "integer",
-                        "description": "Maximum evidence trace batches to read, 1..1000",
-                        "name": "limit",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.SnapshotPreviewResponse"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "404": {
-                        "description": "Not Found",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "405": {
-                        "description": "Method Not Allowed",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/requests/{request_id}/traces": {
-            "get": {
-                "description": "Supports one business request to multiple distributed traces and returns request_id on every trace.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "requests"
-                ],
-                "summary": "List technical traces belonging to a business request",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "BKN request ID",
-                        "name": "request_id",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "type": "integer",
-                        "description": "Page size, 1..200",
-                        "name": "limit",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Opaque pagination cursor",
-                        "name": "cursor",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "Caller-owned conversation ID",
-                        "name": "conversation_id",
-                        "in": "query"
-                    },
-                    {
-                        "type": "string",
-                        "description": "One user interaction ID",
-                        "name": "interaction_id",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.TraceSummaryPage"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "401": {
-                        "description": "Unauthorized",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/traces/{trace_id}/business-graph": {
-            "get": {
-                "description": "Returns claim and business semantic nodes/edges derived from business.refs.resolved events.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "evidence"
-                ],
-                "summary": "Get business semantic graph by trace ID",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "Trace ID",
-                        "name": "trace_id",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "type": "integer",
-                        "description": "Maximum evidence trace batches to read, 1..1000",
-                        "name": "limit",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.BusinessGraphResponse"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "404": {
-                        "description": "Not Found",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "405": {
-                        "description": "Method Not Allowed",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/traces/{trace_id}/evidence-chain": {
-            "get": {
-                "description": "Returns normalized claim, evidence refs, business refs, pagination, partial reasons, and visibility summary for a trace.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "evidence"
-                ],
-                "summary": "Get evidence chain by trace ID",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "Trace ID",
-                        "name": "trace_id",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "type": "integer",
-                        "description": "Maximum evidence trace batches to read, 1..1000",
-                        "name": "limit",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.EvidenceChainResponse"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "404": {
-                        "description": "Not Found",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "405": {
-                        "description": "Method Not Allowed",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
-        "/business-provenance/traces/{trace_id}/snapshot-preview": {
-            "get": {
-                "description": "Returns a metadata-only governed snapshot manifest preview without creating or exposing object storage locations.",
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "evidence"
-                ],
-                "summary": "Get evidence snapshot preview by trace ID",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "Trace ID",
-                        "name": "trace_id",
-                        "in": "path",
-                        "required": true
-                    },
-                    {
-                        "type": "integer",
-                        "description": "Maximum evidence trace batches to read, 1..1000",
-                        "name": "limit",
-                        "in": "query"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/evidencevo.SnapshotPreviewResponse"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "404": {
-                        "description": "Not Found",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "405": {
-                        "description": "Method Not Allowed",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "$ref": "#/definitions/rdto.ErrorResponse"
-                        }
-                    }
-                }
-            }
-        },
         "/conversations": {
             "get": {
                 "produces": [
@@ -3193,26 +2255,6 @@ const docTemplate = `{
                 }
             }
         },
-        "evidencevo.ActionSummary": {
-            "type": "object",
-            "properties": {
-                "approved": {
-                    "type": "integer"
-                },
-                "completed": {
-                    "type": "integer"
-                },
-                "executed": {
-                    "type": "integer"
-                },
-                "last_status": {
-                    "type": "string"
-                },
-                "recommended": {
-                    "type": "integer"
-                }
-            }
-        },
         "evidencevo.ArtifactIngestResponse": {
             "type": "object",
             "properties": {
@@ -3417,112 +2459,6 @@ const docTemplate = `{
                 },
                 "visibility_summary": {
                     "$ref": "#/definitions/evidencevo.VisibilitySummary"
-                }
-            }
-        },
-        "evidencevo.ConversationSummary": {
-            "type": "object",
-            "properties": {
-                "agent_name": {
-                    "type": "string"
-                },
-                "agent_or_app": {
-                    "type": "string"
-                },
-                "application_principal_id": {
-                    "type": "string"
-                },
-                "business_domain": {
-                    "type": "string"
-                },
-                "completed_at": {
-                    "type": "string"
-                },
-                "conversation_id": {
-                    "type": "string"
-                },
-                "duration_ms": {
-                    "type": "integer"
-                },
-                "effective_subject_id": {
-                    "type": "string"
-                },
-                "error_summary": {
-                    "type": "string"
-                },
-                "evidence_completeness": {
-                    "type": "string"
-                },
-                "initiator": {
-                    "type": "string"
-                },
-                "interaction_count": {
-                    "type": "integer"
-                },
-                "knowledge_networks": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "partial_reasons": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "question_preview": {
-                    "type": "string"
-                },
-                "request_count": {
-                    "type": "integer"
-                },
-                "result_preview": {
-                    "type": "string"
-                },
-                "started_at": {
-                    "type": "string"
-                },
-                "status": {
-                    "type": "string"
-                },
-                "trace_count": {
-                    "type": "integer"
-                }
-            }
-        },
-        "evidencevo.ConversationSummaryPage": {
-            "type": "object",
-            "properties": {
-                "entries": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/evidencevo.ConversationSummary"
-                    }
-                },
-                "next_cursor": {
-                    "type": "string"
-                },
-                "page": {
-                    "type": "integer"
-                },
-                "page_size": {
-                    "type": "integer"
-                },
-                "partial": {
-                    "type": "boolean"
-                },
-                "partial_reasons": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "total": {
-                    "type": "integer"
-                },
-                "truncated": {
-                    "type": "boolean"
                 }
             }
         },
@@ -3747,310 +2683,6 @@ const docTemplate = `{
                 },
                 "trace_id": {
                     "type": "string"
-                }
-            }
-        },
-        "evidencevo.InteractionListSummary": {
-            "type": "object",
-            "properties": {
-                "agent_name": {
-                    "type": "string"
-                },
-                "agent_or_app": {
-                    "type": "string"
-                },
-                "application_principal_id": {
-                    "type": "string"
-                },
-                "business_domain": {
-                    "type": "string"
-                },
-                "completed_at": {
-                    "type": "string"
-                },
-                "conversation_id": {
-                    "type": "string"
-                },
-                "duration_ms": {
-                    "type": "integer"
-                },
-                "effective_subject_id": {
-                    "type": "string"
-                },
-                "error_summary": {
-                    "type": "string"
-                },
-                "evidence_completeness": {
-                    "type": "string"
-                },
-                "initiator": {
-                    "type": "string"
-                },
-                "interaction_id": {
-                    "type": "string"
-                },
-                "knowledge_networks": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "partial_reasons": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "question_preview": {
-                    "type": "string"
-                },
-                "request_count": {
-                    "type": "integer"
-                },
-                "result_preview": {
-                    "type": "string"
-                },
-                "started_at": {
-                    "type": "string"
-                },
-                "status": {
-                    "type": "string"
-                },
-                "trace_count": {
-                    "type": "integer"
-                }
-            }
-        },
-        "evidencevo.InteractionSummary": {
-            "type": "object",
-            "properties": {
-                "agent_name": {
-                    "type": "string"
-                },
-                "application_principal_id": {
-                    "type": "string"
-                },
-                "completed_at": {
-                    "type": "string"
-                },
-                "conversation_id": {
-                    "type": "string"
-                },
-                "duration_ms": {
-                    "type": "integer"
-                },
-                "effective_subject_id": {
-                    "type": "string"
-                },
-                "error_summary": {
-                    "type": "string"
-                },
-                "evidence_completeness": {
-                    "type": "string"
-                },
-                "interaction_id": {
-                    "type": "string"
-                },
-                "partial_reasons": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "question_artifact_ref": {
-                    "type": "string"
-                },
-                "question_preview": {
-                    "type": "string"
-                },
-                "requests": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/evidencevo.RequestSummary"
-                    }
-                },
-                "result_artifact_ref": {
-                    "type": "string"
-                },
-                "result_preview": {
-                    "type": "string"
-                },
-                "started_at": {
-                    "type": "string"
-                },
-                "status": {
-                    "type": "string"
-                },
-                "traces": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/evidencevo.TraceSummary"
-                    }
-                }
-            }
-        },
-        "evidencevo.InteractionSummaryPage": {
-            "type": "object",
-            "properties": {
-                "entries": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/evidencevo.InteractionListSummary"
-                    }
-                },
-                "next_cursor": {
-                    "type": "string"
-                },
-                "page": {
-                    "type": "integer"
-                },
-                "page_size": {
-                    "type": "integer"
-                },
-                "partial": {
-                    "type": "boolean"
-                },
-                "partial_reasons": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "total": {
-                    "type": "integer"
-                },
-                "truncated": {
-                    "type": "boolean"
-                }
-            }
-        },
-        "evidencevo.RequestSummary": {
-            "type": "object",
-            "properties": {
-                "action_summary": {
-                    "$ref": "#/definitions/evidencevo.ActionSummary"
-                },
-                "agent_name": {
-                    "type": "string"
-                },
-                "agent_or_app": {
-                    "type": "string"
-                },
-                "application_principal_id": {
-                    "type": "string"
-                },
-                "business_domain": {
-                    "type": "string"
-                },
-                "business_refs": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "completed_at": {
-                    "type": "string"
-                },
-                "controlled_summary": {
-                    "type": "string"
-                },
-                "conversation_id": {
-                    "type": "string"
-                },
-                "duration_ms": {
-                    "type": "integer"
-                },
-                "effective_subject_id": {
-                    "type": "string"
-                },
-                "error_summary": {
-                    "type": "string"
-                },
-                "evidence_completeness": {
-                    "type": "string"
-                },
-                "initiator": {
-                    "type": "string"
-                },
-                "interaction_id": {
-                    "type": "string"
-                },
-                "knowledge_networks": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "operation_id": {
-                    "type": "string"
-                },
-                "operation_key": {
-                    "type": "string"
-                },
-                "partial_reasons": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "question_preview": {
-                    "type": "string"
-                },
-                "request_id": {
-                    "type": "string"
-                },
-                "result_count": {
-                    "type": "integer"
-                },
-                "result_preview": {
-                    "type": "string"
-                },
-                "started_at": {
-                    "type": "string"
-                },
-                "status": {
-                    "type": "string"
-                },
-                "tool_name": {
-                    "type": "string"
-                },
-                "trace_count": {
-                    "type": "integer"
-                }
-            }
-        },
-        "evidencevo.RequestSummaryPage": {
-            "type": "object",
-            "properties": {
-                "entries": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/evidencevo.RequestSummary"
-                    }
-                },
-                "next_cursor": {
-                    "type": "string"
-                },
-                "page": {
-                    "type": "integer"
-                },
-                "page_size": {
-                    "type": "integer"
-                },
-                "partial": {
-                    "type": "boolean"
-                },
-                "partial_reasons": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                },
-                "total": {
-                    "type": "integer"
-                },
-                "truncated": {
-                    "type": "boolean"
                 }
             }
         },
@@ -4989,6 +3621,9 @@ const docTemplate = `{
                     "type": "boolean"
                 },
                 "management_audit": {
+                    "type": "boolean"
+                },
+                "observability_archive_manage": {
                     "type": "boolean"
                 },
                 "security_audit": {
